# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Safalife API is a comprehensive backend service for Safalife's competition management system. This API provides endpoints for health monitoring, user management, competition tracking, and administrative functions. Built with Go using Clean Architecture principles, it offers high performance and reliability for managing large-scale competitions.",
        "title": "Safalife API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Safalife API Support Team",
            "url": "http://www.safalife.com/support",
            "email": "api-support@safalife.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/quran/bookmarks/ayahs/user/{userId}": {
            "get": {
                "description": "Get a paginated list of ayah bookmarks for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "List bookmark ayahs by user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmarks retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.BookmarkAyahListSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quran/last-reads/user/{userId}": {
            "get": {
                "description": "Get a paginated list of last read entries for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "List last reads by user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "relations"
                        ],
                        "type": "string",
                        "example": "relations",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Last reads retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.LastReadListSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
        "/quran/last-reads/user/{userId}/surah/{surahId}": {
            "get": {
                "description": "Get the last read entry for a specific user and surah",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "Get last read by user and surah",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Surah ID",
                        "name": "surahId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "relations"
                        ],
                        "type": "string",
                        "example": "relations",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Last read retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.LastReadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Last read not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
        "/quran/last-reads/{id}": {
            "get": {
                "description": "Get a specific last read entry by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "Get last read by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LastRead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "relations"
                        ],
                        "type": "string",
                        "example": "relations",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Last read retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.LastReadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Last read not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update fields of an existing last read entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "Update last read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LastRead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "LastRead update information",
                        "name": "lastRead",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateLastReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Last read updated successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.LastReadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Last read not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a last read entry by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "Delete last read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LastRead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Last read deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.DeleteLastReadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Last read not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "dto.BookmarkAyahListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BookmarkAyahResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationResponse"
                }
            }
        },
        "dto.BookmarkAyahResponse": {
            "type": "object",
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "dto.LastReadListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.LastReadResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationResponse"
                }
            }
        },
        "dto.LastReadResponse": {
            "type": "object",
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "ayahNumber": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastReadAt": {
                    "type": "string"
                },
                "progressPct": {
                    "type": "number"
                },
                "surahId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateLastReadRequest": {
            "type": "object",
            "required": [
                "ayahId",
                "ayahNumber",
                "id",
                "progressPct",
                "updatedBy"
            ],
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "ayahNumber": {
                    "type": "integer",
                    "minimum": 1
                },
                "id": {
                    "type": "string"
                },
                "progressPct": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "updatedBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "quran.BookmarkAyahListSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.BookmarkAyahListResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Bookmarks retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.DeleteLastReadSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "Last read deleted successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.LastReadListSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.LastReadListResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Last reads retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "quran.LastReadSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.LastReadResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Last read retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.QuranErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/core.ErrorDetail"
                },
                "message": {
                    "type": "string",
                    "example": "Failed to process request"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}